+++ backend/core/detection.py
@@ -12,7 +12,7 @@ from typing import List
 from typing import NoReturn
 from typing import Tuple
 
-import radix
+import network_finder
 import redis
 import yaml
 from kombu import Connection
@@ -341,7 +341,7 @@ class Detection:
             """
             Updates rules everytime it receives a new configuration.
             """
-            self.prefix_tree = radix.Radix()
+            self.prefix_tree = network_finder.NetworkFinder()
             for rule in self.rules:
                 try:
                     for prefix in rule["prefixes"]:
@@ -349,6 +349,8 @@ class Detection:
                             node = self.prefix_tree.search_exact(translated_prefix)
                             if not node:
                                 node = self.prefix_tree.add(translated_prefix)
+                                node.data = {}
+                                node.data["prefix"] = translated_prefix
                                 node.data["confs"] = []
 
                             rule_translated_origin_asn_set = set()
@@ -439,7 +441,7 @@ class Detection:
                 prefix_node = self.prefix_tree.search_best(monitor_event["prefix"])
 
                 if prefix_node:
-                    monitor_event["matched_prefix"] = prefix_node.prefix
+                    monitor_event["matched_prefix"] = prefix_node.data["prefix"]
 
                     try:
                         path_hijacker = -1
@@ -621,7 +623,11 @@ class Detection:
 
         @exception_handler(log)
         def detect_prefix_squatting_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> str:
             """
             Squatting hijack detection.
@@ -634,19 +640,30 @@ class Detection:
 
         @exception_handler(log)
         def detect_prefix_subprefix_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> str:
             """
             Subprefix or exact prefix hijack detection.
             """
             mon_prefix = ipaddress.ip_network(monitor_event["prefix"])
-            if prefix_node.prefixlen < mon_prefix.prefixlen:
+            if (
+                ipaddress.ip_network(prefix_node.data["prefix"]).prefixlen
+                < mon_prefix.prefixlen
+            ):
                 return "S"
             return "E"
 
         @exception_handler(log)
         def detect_path_type_0_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> Tuple[int, str]:
             """
             Origin hijack detection.
@@ -659,7 +676,11 @@ class Detection:
 
         @exception_handler(log)
         def detect_path_type_1_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> Tuple[int, str]:
             """
             Type-1 hijack detection.
@@ -680,42 +701,66 @@ class Detection:
 
         @exception_handler(log)
         def detect_path_type_N_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> Tuple[int, str]:
             # Placeholder for type-N detection (not supported)
             return (-1, "-")
 
         @exception_handler(log)
         def detect_path_type_U_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> Tuple[int, str]:
             # Placeholder for type-U detection (not supported)
             return (-1, "-")
 
         @exception_handler(log)
         def detect_dplane_blackholing_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> str:
             # Placeholder for blackholing detection  (not supported)
             return "-"
 
         @exception_handler(log)
         def detect_dplane_imposture_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> str:
             # Placeholder for imposture detection  (not supported)
             return "-"
 
         @exception_handler(log)
         def detect_dplane_mitm_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> str:
             # Placeholder for mitm detection  (not supported)
             return "-"
 
         @exception_handler(log)
         def detect_pol_leak_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> Tuple[int, str]:
             """
             Route leak hijack detection
@@ -727,7 +772,11 @@ class Detection:
 
         @exception_handler(log)
         def detect_pol_other_hijack(
-            self, monitor_event: Dict, prefix_node: radix.Radix, *args, **kwargs
+            self,
+            monitor_event: Dict,
+            prefix_node: network_finder.NetworkFinder,
+            *args,
+            **kwargs
         ) -> Tuple[int, str]:
             # Placeholder for policy violation detection (not supported)
             return (-1, "-")