+++ backend/core/mitigation.py
@@ -3,7 +3,7 @@ import signal
 import subprocess
 import time
 
-import radix
+import network_finder
 from kombu import Connection
 from kombu import Consumer
 from kombu import Exchange
@@ -155,12 +155,14 @@ class Mitigation:
                     self.init_mitigation()
 
         def init_mitigation(self):
-            self.prefix_tree = radix.Radix()
+            self.prefix_tree = network_finder.NetworkFinder()
             for rule in self.rules:
                 try:
                     for prefix in rule["prefixes"]:
                         for translated_prefix in translate_rfc2622(prefix):
                             node = self.prefix_tree.add(translated_prefix)
+                            node.data = {}
+                            node.data["prefix"] = translated_prefix
                             node.data["mitigation"] = rule["mitigation"]
                 except Exception:
                     log.exception("Exception")